requirements.txt
colorama
ffmpeg-python
librosa
mutagen
pillow